ppi: move data processing to data_processing_utils.ts
